requirements.txt
telethon==1.34.0
python-telegram-bot==20.6
python-dotenv==1.0.0
python-telegram-bot==20.7
python-dotenv==1.0.0
